+++ data-preparation/src/main/kotlin/com/aseemsavio/bibilia/data/BibiliaDataReader.kt
@@ -1,7 +1,6 @@
 package com.aseemsavio.bibilia.data
 
 import kotlinx.coroutines.*
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 import java.net.URL
@@ -13,52 +12,6 @@ val VULGATE_URL = URL("https://raw.githubusercontent.com/aseemsavio/Latin-Vulgat
 private suspend fun getRawBibleJson(url: URL): String =
   withContext(Dispatchers.IO) { url.readText() }
 
-@Serializable
-data class JsonVerse(
-  val chapter: Int,
-  val verse: Int,
-  val textEn: String,
-  val textLa: String,
-  val notes: String? = null
-)
-
-@Serializable
-data class JsonChapter(
-  val chapter: Int,
-  val verses: List<JsonVerse>
-)
-
-@Serializable
-data class JsonBook(
-  val bookNumber: Int,
-  val book: String,
-  val testament: String,
-  val chapters: List<JsonChapter>
-)
-
-@Serializable
-data class JsonTestament(
-  val testament: String,
-  val books: List<JsonBook>
-)
-
-@Serializable
-data class BookNamesItem(
-  val testament: String,
-  val bookNames: Set<String>
-)
-
-@JvmInline
-value class Version(val value: String)
-
-data class VersionInfo(
-  val version: Version,
-  val url: URL
-)
-
-typealias BibleJson = List<JsonBook>
-
-val String.v get() = Version(this)
 
 /**
  * Does map operation in parallel
