+++ database/build.gradle.kts
@@ -7,5 +7,6 @@ plugins {
 
 dependencies {
   implementation(kotlin("stdlib-jdk8"))
+  implementation(project(":domain"))
   implementation(project(":data-preparation"))
 }
